Move insert_member and get_member_by_id to helpers.py
